+++ MSListEmbed/Pages/Index.cshtml.cs
@@ -41,7 +41,7 @@ public class IndexModel : PageModel
             HostingDatesLength = i.Fields!.AdditionalData.TryGetValue("Hosting_x0020_dates_x002f_length", out var dates) ? dates?.ToString() : null,
             LocationDetails = i.Fields!.AdditionalData.TryGetValue("Location_x0020_details_x003a_", out var loc) ? loc?.ToString() : null,
             ImportantDetails = i.Fields!.AdditionalData.TryGetValue("Important_x0020_Details_x003a_", out var important) ? important?.ToString() : null,
-            DisplayOrder = i.Fields!.AdditionalData.TryGetValue("DisplayOrder", out var order) ? order as int? ?? 0 : 0,
+            DisplayOrder = i.Fields!.AdditionalData.TryGetValue("DisplayOrder", out var order) ? order as decimal? ?? 0 : 0,
         }).ToList() ?? [];
     }
 }
